pyproject.toml
[tool.poetry]
name = "repl_python3_RLD"
version = "0.1.0"
description = ""
authors = ["FlorencioOlsen <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
